+++ test/mock_packages/flutter/pubspec.yaml
@@ -1,0 +1,4 @@
+name: flutter
+
+environment:
+  sdk: '>=2.2.2 <3.0.0'
