+++ test/rules/.mock_packages
@@ -1,1 +1,2 @@
 fixnum:../mock_packages/fixnum/lib/
+flutter:../mock_packages/flutter/lib/
